+++ rdb/dao_struct.go
@@ -7,18 +7,26 @@ type EmployeePlural struct {
 }
 
 type Employee struct {
-	Id       int    `json:"id"`
-	Name     string `json:"name"`
-	Email    string `json:"email"`
-	Position string `json:"position"`
-	Account  string `json:"account"`
-	Payroll  int    `json:"payroll"`
-	Currency string `json:"curr"`
-	Date     int    `json:"date"`
+	Id       int     `json:"key"`
+	Name     string  `json:"name"`
+	Email    string  `json:"email"`
+	Position string  `json:"position"`
+	Account  string  `json:"account"`
+	Payroll  int     `json:"payroll"`
+	Currency string  `json:"curr"`
+	Date     int     `json:"date"`
+	TimeLeft float64 `json:"secLeft"`
 }
 
 type EmployeeOnboard struct {
 	MyEmployerId   string
 	MyEmployeeId   int
 	EmploymentType string
+}
+
+type Payment struct {
+	Name     string `json:"name"`
+	Payroll  int    `json:"payroll"`
+	Currency string `json:"curr"`
+	Time     string `json:"time"`
 }